xiaofa-bot/v2/requirements.txt
python-telegram-bot==21.0.1
anthropic==0.40.0
notion-client==2.2.1
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
